[refactor](#15) UserAccuracyDtoFactory 클래스 생성 -> dto빌드 메서드 위치 변경
