+++ src/main/java/com/poweranger/hai_duo/user/api/factory/UserFactory.java
@@ -21,4 +21,5 @@ public class UserFactory {
                 .lastAccessedAt(LocalDateTime.now())
                 .build();
     }
+
 }
