+++ src/main/java/com/poweranger/hai_duo/user/application/service/UserAccuracyService.java
@@ -6,6 +6,8 @@ import com.poweranger.hai_duo.progress.domain.repository.StageRepository;
 import com.poweranger.hai_duo.user.api.dto.UserAccuracyByChapterDto;
 import com.poweranger.hai_duo.user.api.dto.UserAccuracyByStageDto;
 import com.poweranger.hai_duo.user.api.dto.UserAccuracyDto;
+import com.poweranger.hai_duo.user.api.factory.UserAccuracyDtoFactory;
+import com.poweranger.hai_duo.user.api.factory.UserFactory;
 import lombok.RequiredArgsConstructor;
 import org.bson.Document;
 import org.springframework.data.mongodb.core.MongoTemplate;
@@ -20,17 +22,18 @@ public class UserAccuracyService {
 
     private final MongoTemplate mongoTemplate;
     private final StageRepository stageRepository;
+    private final UserAccuracyDtoFactory userAccuracyDtoFactory;
 
     public UserAccuracyDto getUserAccuracy(Long userId) {
         Aggregation aggregation = buildUserAggregation(userId);
         Document result = executeAggregation(aggregation);
-        return buildUserAccuracyDto(result);
+        return userAccuracyDtoFactory.buildUserAccuracyDto(result);
     }
 
     public UserAccuracyByStageDto getUserAccuracyByStage(Long userId, Long stageId) {
         Aggregation aggregation = buildStageAggregation(userId, stageId);
         Document result = executeAggregation(aggregation);
-        return buildUserAccuracyDtoByStage(result);
+        return userAccuracyDtoFactory.buildUserAccuracyDtoByStage(result);
     }
 
     public UserAccuracyByChapterDto getUserAccuracyByChapter(Long userId, Long chapterId) {
@@ -39,7 +42,7 @@ public class UserAccuracyService {
 
         Aggregation aggregation = buildChapterAggregation(userId, stageIds);
         Document result = executeAggregation(aggregation);
-        return buildUserAccuracyDtoByChapter(result, chapterId);
+        return userAccuracyDtoFactory.buildUserAccuracyDtoByChapter(result, chapterId);
     }
 
     // --- Aggregation Builders ---
@@ -90,37 +93,6 @@ public class UserAccuracyService {
         return result;
     }
 
-    // --- DTO Builders ---
-
-    private UserAccuracyDto buildUserAccuracyDto(Document result) {
-        return new UserAccuracyDto(
-                convertToLong(result.get("userId")),
-                convertToInt(result.get("totalCount")),
-                convertToInt(result.get("correctCount")),
-                calculateAccuracyRate(result)
-        );
-    }
-
-    private UserAccuracyByStageDto buildUserAccuracyDtoByStage(Document result) {
-        return new UserAccuracyByStageDto(
-                convertToLong(result.get("userId")),
-                convertToLong(result.get("stageId")),
-                convertToInt(result.get("totalCount")),
-                convertToInt(result.get("correctCount")),
-                calculateAccuracyRate(result)
-        );
-    }
-
-    private UserAccuracyByChapterDto buildUserAccuracyDtoByChapter(Document result, Long chapterId) {
-        return new UserAccuracyByChapterDto(
-                convertToLong(result.get("userId")),
-                chapterId,
-                convertToInt(result.get("totalCount")),
-                convertToInt(result.get("correctCount")),
-                calculateAccuracyRate(result)
-        );
-    }
-
     // --- Validation ---
 
     private void validateAggregationResult(Document result) {
@@ -133,28 +105,5 @@ public class UserAccuracyService {
         if (stageIds == null || stageIds.isEmpty()) {
             throw new GeneralException(ErrorStatus.USER_ACCURACY_NOT_FOUND);
         }
-    }
-
-    // --- Converter / Calculator ---
-
-    private Long convertToLong(Object value) {
-        if (value instanceof Number number) {
-            return number.longValue();
-        }
-        throw new GeneralException(ErrorStatus.USER_ACCURACY_NOT_FOUND);
-    }
-
-    private int convertToInt(Object value) {
-        if (value instanceof Number number) {
-            return number.intValue();
-        }
-        return 0;
-    }
-
-    private float calculateAccuracyRate(Document result) {
-        int total = convertToInt(result.get("totalCount"));
-        int correct = convertToInt(result.get("correctCount"));
-        if (total == 0) return 0.0f;
-        return Math.round(((float) correct / total) * 10000) / 100.0f;
     }
 }
